+++ client/src/pages/Signup/Signup.jsx
@@ -7,6 +7,7 @@ import Button from '../../components/common/Button/Button';
 import Oauth from '../../components/common/Oauth/Oauth';
 import { Link, useNavigate } from 'react-router-dom';
 import axios from 'axios';
+import { validate } from '../../utils/validate-signup';
 
 export default function Signup() {
   const navigate = useNavigate();
@@ -33,80 +34,6 @@ export default function Signup() {
     password: false,
     password_confirm: false,
   });
-  console.log(valid);
-  console.log(form);
-
-  const validate = (form) => {
-    const newErrors = {
-      name: '',
-      phone: '',
-      email: '',
-      password: '',
-      password_confirm: '',
-    };
-    if (!/^[\uac00-\ud7a3]+$/.test(form.name)) {
-      newErrors.name = '이름은 한글만 입력 가능합니다.';
-      setValid((prevValid) => ({ ...prevValid, name: false }));
-    } else if (!/^(?=.{2,20}$)[가-힣]{2,20}$/.test(form.name)) {
-      newErrors.name = '이름은 2~20자 사이의 한글만 입력 가능합니다.';
-      setValid((prevValid) => ({ ...prevValid, name: false }));
-    } else {
-      newErrors.name = '올바른 형식입니다.';
-      setValid((prevValid) => ({ ...prevValid, name: true }));
-    }
-
-    if (form.phone.trim() === '') {
-      newErrors.phone = '휴대폰 번호를 입력해주세요.';
-      setValid((prevValid) => ({ ...prevValid, phone: false }));
-    } else if (!/^010-\d{4}-\d{4}$/.test(form.phone)) {
-      newErrors.phone =
-        "휴대폰 번호는 010으로 시작하는 11자리 숫자와 '-'로 구성되어야 합니다.";
-      setValid((prevValid) => ({ ...prevValid, phone: false }));
-    } else {
-      newErrors.phone = '올바른 형식입니다.';
-      setValid((prevValid) => ({ ...prevValid, phone: true }));
-    }
-
-    if (form.email.trim() === '') {
-      newErrors.email = '이메일을 입력해주세요.';
-      setValid((prevValid) => ({ ...prevValid, email: false }));
-    } else if (!/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i.test(form.email)) {
-      newErrors.email = '올바른 이메일 형식이 아닙니다.';
-      setValid((prevValid) => ({ ...prevValid, email: false }));
-    } else {
-      newErrors.email = '올바른 형식입니다.';
-      setValid((prevValid) => ({ ...prevValid, email: true }));
-    }
-
-    if (form.password.trim() === '') {
-      newErrors.password = '비밀번호를 입력해주세요.';
-      setValid((prevValid) => ({ ...prevValid, password: false }));
-    } else if (form.password.length < 8) {
-      newErrors.password = '비밀번호는 8자리 이상이어야 합니다.';
-      setValid((prevValid) => ({ ...prevValid, password: false }));
-    } else if (
-      !/^(?=.*[a-zA-Z])(?=.*[0-9])(?=.*[!@#$%^&*]).{8,}$/.test(form.password)
-    ) {
-      newErrors.password =
-        '영문자, 숫자, 특수문자를 포함한 8글자 이상의 비밀번호를 입력해주세요.';
-      setValid((prevValid) => ({ ...prevValid, password: false }));
-    } else {
-      newErrors.password = '올바른 형식입니다.';
-      setValid((prevValid) => ({ ...prevValid, password: true }));
-    }
-
-    if (form.password_confirm.trim() === '') {
-      newErrors.password_confirm = '비밀번호를 다시 입력해주세요.';
-      setValid((prevValid) => ({ ...prevValid, password_confirm: false }));
-    } else if (form.password !== form.password_confirm) {
-      newErrors.password_confirm = '비밀번호가 일치하지 않습니다.';
-      setValid((prevValid) => ({ ...prevValid, password_confirm: false }));
-    } else if (form.password === form.password_confirm) {
-      newErrors.password_confirm = '비밀번호가 일치합니다.';
-      setValid((prevValid) => ({ ...prevValid, password_confirm: true }));
-    }
-    setErrors(newErrors);
-  };
 
   const allTrue = Object.values(valid).every((value) => value === true);
 
@@ -139,10 +66,13 @@ export default function Signup() {
       e.preventDefault(); // 입력 막기
       return;
     }
+    console.log({ ...form, [name]: value });
     setForm((prevForm) => ({ ...prevForm, [name]: value }));
     // 문제발생: 최신 form value를 전달받지 못함(setState가 비동기로 동작)
     // 해결방법: form value를 업데이트해서 전달
-    validate({ ...form, [name]: value });
+    const [newErrors, newValid] = validate({ ...form, [name]: value });
+    setErrors(newErrors);
+    setValid(newValid);
   };
 
   return (
